+++ core/src/main/java/io/github/wysohn/triggerreactor/core/manager/AbstractJavascriptBasedManager.java
@@ -17,6 +17,7 @@
 package io.github.wysohn.triggerreactor.core.manager;
 
 import io.github.wysohn.triggerreactor.core.main.TriggerReactorCore;
+import io.github.wysohn.triggerreactor.core.manager.evaluable.IEvaluable;
 
 import javax.script.Bindings;
 import javax.script.ScriptContext;
