+++ core/src/main/java/io/github/wysohn/triggerreactor/core/manager/ExecutorManager.java
@@ -17,6 +17,7 @@
 package io.github.wysohn.triggerreactor.core.manager;
 
 import io.github.wysohn.triggerreactor.core.main.TriggerReactorCore;
+import io.github.wysohn.triggerreactor.core.manager.evaluable.JSExecutor;
 import io.github.wysohn.triggerreactor.core.script.interpreter.Executor;
 import io.github.wysohn.triggerreactor.tools.JarUtil;
