+++ core/src/test/java/io/github/wysohn/triggerreactor/core/manager/trigger/area/AreaTriggerManagerTest.java
@@ -1,9 +1,26 @@
+/*
+ * Copyright (C) 2022. TriggerReactor Team
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package io.github.wysohn.triggerreactor.core.manager.trigger.area;
 
 import io.github.wysohn.triggerreactor.core.bridge.entity.IPlayer;
 import io.github.wysohn.triggerreactor.core.config.InvalidTrgConfigurationException;
+import io.github.wysohn.triggerreactor.core.main.IGameManagement;
 import io.github.wysohn.triggerreactor.core.main.TriggerReactorCore;
-import io.github.wysohn.triggerreactor.core.manager.IGameStateSupervisor;
 import io.github.wysohn.triggerreactor.core.manager.location.Area;
 import io.github.wysohn.triggerreactor.core.manager.location.SimpleLocation;
 import io.github.wysohn.triggerreactor.core.manager.trigger.ITriggerLoader;
@@ -22,7 +39,7 @@ public class AreaTriggerManagerTest {
 
     TriggerReactorCore core;
     TaskSupervisor taskSupervisor;
-    IGameStateSupervisor gameStateSupervisor;
+    IGameManagement gameManagement;
     ITriggerLoader<AreaTrigger> loader;
 
     AreaTriggerManager manager;
@@ -35,10 +52,10 @@ public class AreaTriggerManagerTest {
         instanceField.set(null, core);
 
         taskSupervisor = mock(TaskSupervisor.class);
-        gameStateSupervisor = mock(IGameStateSupervisor.class);
+        gameManagement = mock(IGameManagement.class);
         loader = mock(ITriggerLoader.class);
 
-        manager = new AreaTriggerManager(core, taskSupervisor, gameStateSupervisor, loader);
+        manager = new AreaTriggerManager(core, taskSupervisor, gameManagement, loader);
     }
 
     @Test
